author create: found a way to correct the date input
date input is local to the client, yet it gets parsed as utc/zulu
date. To correct this, a hidden input field containing the client's
timezone offset is forwarded to the server/controller on submit,
which is used to adjust the date to the correct utc date on
sanitization. From that point, normal validation and sanitization
chaining can proceed